[IMP] l10n_co: support consumption taxes for Colombia by default
Consumption taxes are required for alcohol and other products, with
the POS module for Colombia coming out soon it's necessary to support
such taxes out of the box for when it's released. A migration script has
been added to automatically load these new taxes and accounts that will
be necessary.

task-4634313

closes odoo/odoo#210715

X-original-commit: 868aeba46425f9091f9e9779a46cdea1d957bd8f
Related: odoo/enterprise#85930
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Andrew Gavgavian (andg) <andg@odoo.com>